refactor(seo): split seo.ts into seo.data.ts + seo.ts to comply with 300-line limit
- Move all types, constants, and data arrays to src/lib/seo.data.ts (256 lines)
- Keep utility functions in src/lib/seo.ts (178 lines) with re-exports for
  backward-compatible import surface
- Add 30 unit tests in src/lib/__tests__/seo.test.ts covering absoluteUrl,
  findLocalGrowthPage, localGrowthMetadata, localBusinessSchema, serviceSchema,
  faqSchema, breadcrumbSchema, localGrowthSchemas breadcrumb depth, and
  localGrowthPages data integrity

All existing imports from @/lib/seo remain stable.
